Изменены названия некоторых методов, отрисовка интерфейса вынесена в отдельный файл, расставлены марки
