+++ Trainee app/BasicNoteScreen/NoteViewController.swift
@@ -8,70 +8,19 @@
 import UIKit
 
 final class NoteViewController: UIViewController, UITextFieldDelegate {
-    struct Constants {
-        static let titleFont: UIFont = .systemFont(ofSize: 22, weight: .bold)
-        static let noteFont: UIFont = .systemFont(ofSize: 14, weight: .regular)
-        static let navBarTitle: String = "NotePad"
-        static let cornerRadius: CGFloat = 5
-    }
 
-    lazy var dateField: UITextField = {
-        let field = UITextField()
-        field.font = Constants.noteFont
-        field.sizeToFit()
-        field.backgroundColor = .systemGray3
-        field.layer.cornerCurve = .circular
-        field.layer.cornerRadius = Constants.cornerRadius
-        field.placeholder = Date().toString(format: "Дата: dd MMMM yyyy")
-        field.inputView = datePicker
-        field.translatesAutoresizingMaskIntoConstraints = false
-        return field
-    }()
+    let noteView = NoteView(frame: .zero)
 
-    lazy var datePicker: UIDatePicker = {
-        let date = UIDatePicker()
-        date.datePickerMode = .date
-        date.preferredDatePickerStyle = .wheels
-        date.addTarget(
-            self,
-            action: #selector(self.dateChanged),
-            for: .allEvents
-        )
-        date.translatesAutoresizingMaskIntoConstraints = false
-        return date
-    }()
-
-    lazy var titleField: UITextField = {
-        let field = UITextField()
-        field.placeholder = "Name your note..."
-        field.sizeToFit()
-        field.textColor = UIColor.white
-        field.font = Constants.titleFont
-        field.translatesAutoresizingMaskIntoConstraints = false
-        return field
-    }()
-
-    lazy var noteText: UITextView = {
-        let text = UITextView()
-        text.backgroundColor = .systemGray3
-        text.layer.cornerCurve = .circular
-        text.layer.cornerRadius = Constants.cornerRadius
-        text.sizeToFit()
-        text.isEditable = true
-        text.textColor = UIColor.white
-        text.font = Constants.noteFont
-        text.translatesAutoresizingMaskIntoConstraints = false
-        return text
-    }()
-// MARK: ViewDidLoad
+    // MARK: ViewDidLoad
     override func viewDidLoad() {
         super.viewDidLoad()
-        titleField.delegate = self
-        noteText.becomeFirstResponder()
-        setupMainView()
+        noteView.titleField.delegate = self
+        noteView.noteText.becomeFirstResponder()
+        getViewData()
         setupNavBar()
+        view.addSubview(noteView)
     }
-// MARK: Methods
+    // MARK: Methods
     override func touchesBegan(_ touches: Set<UITouch>, with event: UIEvent?) {
         view.endEditing(true)
         super.touchesBegan(
@@ -80,50 +29,26 @@ final class NoteViewController: UIViewController, UITextFieldDelegate {
         )
     }
 
-    func setupMainView() {
-        getViewData()
-        view.addSubview(dateField)
-        view.addSubview(noteText)
-        view.addSubview(titleField)
-        view.backgroundColor = .systemGray4
-        title = Constants.navBarTitle
-
-        NSLayoutConstraint.activate([
-            titleField.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor, constant: 20),
-            titleField.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor, constant: 20),
-            titleField.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor, constant: -20),
-            noteText.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor, constant: 20),
-            noteText.topAnchor.constraint(equalTo: dateField.bottomAnchor, constant: 5),
-            noteText.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor, constant: -20),
-            noteText.bottomAnchor.constraint(equalTo: view.safeAreaLayoutGuide.bottomAnchor),
-            dateField.topAnchor.constraint(equalTo: titleField.bottomAnchor, constant: 10),
-            dateField.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor, constant: 20),
-            dateField.heightAnchor.constraint(equalToConstant: 28)
-        ])
-    }
-
-    @objc func dateChanged() {
-        dateField.text = "\(datePicker.date.toString(format: "Дата: dd MMMM yyyy"))"
+    override func viewWillLayoutSubviews() {
+        noteView.frame = CGRect(x: 0, y: 0, width: view.frame.size.width, height: view.frame.size.height)
     }
 
     @objc func saveViewData() {
-        resignResponders()
-        let model = NoteModel(title: titleField.text, noteText: noteText.text, date: dateField.text)
-        model.checkEmptyNoteAndAlert(model: model, rootVC: self)
-    }
-
-    func resignResponders() {
-        noteText.resignFirstResponder()
-        titleField.resignFirstResponder()
-        dateField.resignFirstResponder()
+        noteView.resignResponders()
+        let model = NoteModel(
+            title: noteView.titleField.text,
+            noteText: noteView.noteText.text,
+            date: noteView.dateField.text
+        )
+        model.saveNoteOrAlert(model: model, rootVC: self)
     }
 
     func getViewData() {
         if let decodedNote = UserDefaults.standard.object(forKey: "first") as? Data {
             if let noteData = try? JSONDecoder().decode(NoteModel.self, from: decodedNote) {
-                titleField.text = noteData.title
-                noteText.text = noteData.noteText
-                dateField.text = noteData.date
+                noteView.titleField.text = noteData.title
+                noteView.noteText.text = noteData.noteText
+                noteView.dateField.text = noteData.date
             }
         }
     }
@@ -136,21 +61,6 @@ final class NoteViewController: UIViewController, UITextFieldDelegate {
             action: #selector(saveViewData)
         )
         navigationItem.rightBarButtonItem = saveButton
+        title = "Note Pad"
     }
 }
-// MARK: Extensions
-extension Date {
-    func toString(format: String) -> String {
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = format
-        dateFormatter.locale = Locale(identifier: "ru")
-        return dateFormatter.string(from: self)
-    }
-}
-
-extension NoteViewController: UITextViewDelegate {
-    func textFieldShouldReturn(_ textField: UITextField) -> Bool {
-        titleField.resignFirstResponder()
-        return true
-    }
- }
